+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -94,7 +94,7 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-28T06:00:16.795Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-28T07:27:15.078Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")