+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -321,7 +321,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSPJQ3F31TJ453SWDGZ08MW5"
+        "01KSPQPAA67P5KT2AJRCEYZXDZ"
     )
     r"""The unique identifier of the data source"""