+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSPJQ3F1SSAMMMNBJ56QCC7S"
+        "01KSPQPAA3FHWPR3DX4RD79FGE"
     )
     r"""The unique identifier of the data source"""