+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSPJQ3F2095N3QJR0Q8ABS5M"
+        "01KSPQPAA5Y4RDJNJ5DVB6NFYV"
     )
     r"""The unique identifier of the data source"""